+++ src/ui.rs
@@ -1,3 +1,4 @@
 pub mod config;
+mod popup_state;
 pub mod view;
 mod view_state;
